Change position of Advanced Params Setting
